test_syntax.py
#!/usr/bin/env python3

# Test script to isolate syntax issues

try:
    import psycopg
    print("psycopg import successful")
except ImportError as e:
    print(f"Import error: {e}")

# Test the problematic code structure
def test_connection_logic():
    """Test the connection logic structure"""
    try:
        # Simulate the connection test
        print("Testing connection logic...")
        return True
    except:
        print("Connection test failed")
        pass

    # Create new connection
    try:
        print("Creating new connection...")
        return True
    except Exception as e:
        print(f"Connection error: {e}")
        raise e

if __name__ == "__main__":
    print("Testing syntax...")
    result = test_connection_logic()
    print(f"Test result: {result}")
